Prevent removing temp folders in macOS + Python 3

+++ green/process.py
@@ -254,7 +254,9 @@ def poolRunner(target, queue, coverage_number=None, omit_patterns=[]):  # pragma
 
     def cleanup():
         # Restore the state of the temp directory
-        shutil.rmtree(tempfile.tempdir, ignore_errors=True)
+        # TODO: Make this not necessary on macOS+Python3 (see #173)
+        if sys.platform != 'darwin' or sys.version_info[0] == 2:
+            shutil.rmtree(tempfile.tempdir, ignore_errors=True)
         tempfile.tempdir = saved_tempdir
         queue.put(None)
         # Finish coverage
